move sortlgc to common
