+++ common/sortlgc/genzfunc.go
