+++ core/blockchain_viction.go
@@ -24,10 +24,10 @@ import (
 	"sort"
 
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/sortlgc"
 	"github.com/ethereum/go-ethereum/consensus/posv"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/sortlgc"
 )
 
 // Commit native trading/lending trie nodes for the given block to their LevelDB backing stores.
